Cast to int wc jar count
While using Kyuubi on my local environment (OSX) I'm having issues with
`wc` returning the result with a preffixed tab character.

Due to this issue, it fails to compare "	1" with "1" and script
is aborted due to duplicated JARs found.

To solve that, I'm casting `wc` result to integer and perform integer
comparisions, which seems to solve the issue.