Add Customized Configuration Item: kyuubi.statement.id
For broadcast, Spark will introduce a new runId as SPARK_JOB_GROUP_ID. That cause we can't follow up this statement's whole life by SPARK_JOB_GROUP_ID.
This configuration item will be set when this statement was submitted and it will not be changed.
You can get the dependency relationship from job or stage to statement by this configuration item.