[KYUUBI #820] [KYUUBI 817] Fix KyuubiDriverSuite UT failure because of NPE
### _Why are the changes needed?_

See issue #817 , when we pull the project and do not invoke any build action, e.g. `mvn package` or `mvn generate-resources`, we would miss the `kyuubi-version-info.properties` file, it caused the NPE.

As discussed under issue #817 , this PR tries to check it and enhance the exception message to make it more friendly to newbie developers.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.readthedocs.io/en/latest/tools/testing.html#running-tests) locally before make a pull request

- [ ] Verified in the local env and no need to add test cases.

Closes #820 from yanghua/KYUUBI-817.

Closes #820

9464d07b [yanghua] [KYUUBI 817] Fix KyuubiDriverSuite UT failure because of missing kyuubi-version-info.properties

Authored-by: yanghua <yanghua1127@gmail.com>
Signed-off-by: Kent Yao <yao@apache.org>