[KYUUBI #2214] Add Spark Engine on Kubernetes integration test
### _Why are the changes needed?_

Rename `kyuubi-deployment-kubernetes-it` to `kyuubi-kubernetes-it` which include:
- Kyuubi Server on Kubernetes it
- Kyuubi Spark Engine on Kubernetes it

Now, we use the package name to distingush the scope of it:
- deployment package is for Kyuubi Server
- spark package is for Spark Engine

So it's easy to add Flink Engine on kubernetes it in future.

Note that, this PR only supports Spark client mode on Kubernetes it.

Since Spark only have two tags of docker image https://hub.docker.com/r/apache/spark/tags, this PR only test 3.2.1 version

### _How was this patch tested?_
- [x] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2214 from ulysses-you/spark-on-k8s-client.

Closes #2214

f6dda432 [ulysses-you] Add Spark Engine on Kubernetes integration test

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: ulysses-you <ulyssesyou@apache.org>