[KYUUBI #1936][FOLLOWUP] Stop updating credentials when credentials are expired
### _Why are the changes needed?_

This is the follow-up for the comments in https://github.com/apache/incubator-kyuubi/pull/2210

------
But considering the following scenario, I think we should also check whether userRef in `userCredentialsRefMap` has been replaced by a new one .

`userRef` was waiting for next renewal.
`userRef` expired and was removed from `userCredentialsRefMap`
A new `userRef` was created because of user visiting.
Renewal of old userRef started again.
If we do not check, both old and new userRef will renew periodically.

modified the code and added test case for the scenario.

### _How was this patch tested?_
- [x] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2459 from lightning-L/kyuubi-1936.

Closes #1936

0cef07f0 [Tianlin Liao] [KYUUBI #1936][FOLLOWUP] stop updating credentials when credentials are expired

Authored-by: Tianlin Liao <tiliao@ebay.com>
Signed-off-by: Fei Wang <fwang12@ebay.com>