[KYUUBI #2582] Minimize Travis build and test
### _Why are the changes needed?_

The current Travis building is not reliable for the reason of resource limits and maybe some other restrictions.

Because the server module has a lot of tests for different engines, spark, hive, etc, we run a lot of duplicated tests on Travis.

In the PR, I change the jobs to run the tests with `-DwildcardSuites`, only those related to the corresponding module will be run in a particular job.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2582 from yaooqinn/travis2.

Closes #2582

acf69536 [Kent Yao] Minimize Travis build and test
dd7072ad [Kent Yao] Minimize Travis build and test
22acfba5 [Kent Yao] Minimize Travis build and test

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>