[KYUUBI #2642] Fix flaky test - JpsApplicationOperation with spark local mode
### _Why are the changes needed?_

To close #2642

In this PR, we put the `getApplicationInfoByTag` and `killApplicationByTag` into the same `eventually` block to try to fix the confused flaky test.
Since we kill the application with signal `SIGTERM`,  so the process is not terminated immediately.
[SIGTERM vs SIGKILL: What's the Difference?](https://linuxhandbook.com/sigterm-vs-sigkill/)
So we have to wrap the following `getApplicationInfoByTag` block(assert the application is finished) with `eventually` after `killApplicationByTag`.

BTW, we fix a potential issue in case that the `jps` command is not in the PATH.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2669 from turboFei/fix_flaky_test.

Closes #2642

14d2ddfb [Fei Wang] revert
1d8ec692 [Fei Wang] add eventually
9a89609b [Fei Wang] move to eventually and have a try
3cec9063 [Fei Wang] remove catch block
a6a8633c [Fei Wang] fix flaky test

Authored-by: Fei Wang <fwang12@ebay.com>
Signed-off-by: Fei Wang <fwang12@ebay.com>