[KYUUBI #2978] [SUB-TASK][KPIP-4] If batch app status not found from cluster manager, fall back to metadata store
### _Why are the changes needed?_

Because the kyuubi batch is not closed once the session completed(reserve the batch log for sometime).

If a batch has been finished for some time, it might can not be found from RM, which has max reserved application list limitation, for this case, we need fall back to metadata for batch info response.

Before:
<img width="1085" alt="image" src="https://user-images.githubusercontent.com/6757692/176682626-294da2bd-eb37-4b6f-95e9-052990fa2a85.png">

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [x] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #2978 from turboFei/fall_back_metadata.

Closes #2978

a0a51ac0 [Fei Wang] If batch app status not found, fall back to metadata

Authored-by: Fei Wang <fwang12@ebay.com>
Signed-off-by: Fei Wang <fwang12@ebay.com>