[KYUUBI #3715] Remove unexpected mysql connector jar
Fix #3715.

It was added in https://github.com/apache/incubator-kyuubi/pull/3240, and I didn't notice a binary Jar was shipped in the patch, and the worse thing is it was included in the source release tarball.

- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #3716 from pan3793/mysql.

Closes #3715

0fdcac98 [Cheng Pan] fix
833d4679 [Cheng Pan] nit
5d098397 [Cheng Pan] [KYUUBI #3715] Remove unexpected mysql connector jar

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Cheng Pan <chengpan@apache.org>
(cherry picked from commit 689cbe2c18613aca43d16aa2db4f5c76b711c706)
Signed-off-by: Cheng Pan <chengpan@apache.org>