[KYUUBI #4164] Accumulate the operation terminal state
### _Why are the changes needed?_

For the KyuubiOperation, its final state should be `CLOSED`, and now the `FINISHED`, `ERROR` states, which are important for kyuubi administer, are not accumulated.

In this pr, we check whether the old state is terminal state, if it is, do not decrease it.

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.apache.org/docs/latest/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4164 from turboFei/terminate_acc.

Closes #4164

16e559c6f [fwang12] comment
c9d86f4b6 [fwang12] add ut
4fcf6accc [fwang12] Accumulate the terminate state

Authored-by: fwang12 <fwang12@ebay.com>
Signed-off-by: fwang12 <fwang12@ebay.com>