[KYUUBI #4452] Strip the redundant leading and tailing slash of getZooKeeperNamespace.
### _Why are the changes needed?_

As described in the discussion:  <https://github.com/apache/kyuubi/discussions/4436>

When the zk namespace contains leading and trailing slashes, the exception will cause the client connection to fail, and the exception prompt information is not clear.

So we strip the redundant leading and tailing slashes of getZooKeeperNamespace and add exception stack information at the same time.

### _How was this patch tested?_
- [x] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [x] [Run test](https://kyuubi.readthedocs.io/en/master/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4454 from merrily01/zk-namespace.

Closes #4452

62290545e [maruilei] Address comments.
cee43cd61 [maruilei] [KYUUBI #4452] Strip the redundant leading and tailing slash of getZooKeeperNamespace.

Authored-by: maruilei <maruilei@58.com>
Signed-off-by: Cheng Pan <chengpan@apache.org>